+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v5.10.0 (2025-12-27)
+
+### Feat
+
+- add get_role_composites_by_id method (#680)
+
+## v5.9.0 (2025-12-27)
+
+### Feat
+
+- add pool_maxsize parameter to connection managers (#651)
+
 ## v5.8.1 (2025-08-19)
 
 ### Fix